feat: add PostComact component
# Manual description
Plan to add a PostCard component  which is different shape of view for posts.

# Auto description
This commit introduces the PostCompact component to present blog post details in a compact, collapsible format. The Blog page is updated to utilize this new component, displaying a list of dummy blog posts with tags and separators.